+++ packages/http_client_cache/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1
 
 - add `noCache` and `noStore` boolean parameters to `CacheControl` factory constructors to preserve important cache directives when transforming response headers
+- update libs and protobuf Dart files
 
 ## 1.0.0